# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API do Sistema de Suporte TI
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Listar usuários (ADMIN)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, ACTIVE, INACTIVE]
        - name: role
          in: query
          schema:
            type: string
            enum: [USER, COORDINATOR, ANALYST, ADMIN]
      responses:
        "200":
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Obter usuário por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Atualizar usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/approve:
    post:
      operationId: approveUser
      tags: [users]
      summary: Aprovar usuário pendente (ADMIN)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveUserBody"
      responses:
        "200":
          description: Usuário aprovado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ───────── TICKETS ─────────
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: Listar chamados
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        - name: type
          in: query
          schema:
            type: string
            enum: [SOFTWARE, HARDWARE]
        - name: priority
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH]
        - name: uf
          in: query
          schema:
            type: string
        - name: municipality
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de chamados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Abrir novo chamado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketBody"
      responses:
        "201":
          description: Chamado criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Obter chamado por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chamado encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "404":
          description: Chamado não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Atualizar chamado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketBody"
      responses:
        "200":
          description: Chamado atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Chamado não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tickets/{id}/assign:
    post:
      operationId: assignTicket
      tags: [tickets]
      summary: Atribuir chamado a analista
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTicketBody"
      responses:
        "200":
          description: Chamado atribuído
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Chamado não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ───────── MESSAGES ─────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, COORDINATOR, ANALYST, ADMIN]
        status:
          type: string
          enum: [PENDING, ACTIVE, INACTIVE]
        uf:
          type: string
        municipality:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - status
        - uf
        - municipality
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        uf:
          type: string
        municipality:
          type: string
    ApproveUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [USER, COORDINATOR, ANALYST, ADMIN]
      required:
        - role

    # ─── Ticket ───
    Ticket:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [SOFTWARE, HARDWARE]
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        uf:
          type: string
        municipality:
          type: string
        createdById:
          type: integer
        assignedToId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          $ref: "#/components/schemas/UserRef"
        assignedTo:
          oneOf:
            - $ref: "#/components/schemas/UserRef"
            - type: "null"
      required:
        - id
        - title
        - description
        - type
        - status
        - priority
        - uf
        - municipality
        - createdById
        - assignedToId
        - createdAt
        - updatedAt
        - createdBy
    TicketDetail:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
            rating:
              oneOf:
                - $ref: "#/components/schemas/Rating"
                - type: "null"
    UserRef:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, COORDINATOR, ANALYST, ADMIN]
      required:
        - id
        - name
        - email
        - role
    CreateTicketBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [SOFTWARE, HARDWARE]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
      required:
        - title
        - description
        - type
        - priority
    UpdateTicketBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
    AssignTicketBody:
      type: object
      properties:
        assignedToId:
          type: integer
      required:
        - assignedToId

    # ─── Message ───
    Message:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        senderId:
          type: integer
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        sender:
          $ref: "#/components/schemas/UserRef"
      required:
        - id
        - ticketId
        - senderId
        - message
        - createdAt
        - sender
...
    Rating:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        userId:
          type: integer
        score:
          type: integer
          minimum: 1
          maximum: 5
        feedback:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ticketId
        - userId
        - score
        - createdAt
...