# lib/api-zod/src/generated/types/ticketDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API do Sistema de Suporte TI
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";
import type { Rating } from "./rating";
import type { Ticket } from "./ticket";

export type TicketDetail = Ticket & {
  messages?: Message[];
  rating?: Rating | null;
};